# models/apiDocumentation/openapi.yaml
openapi: 3.0.0
info:
  title: SFL
  contact: {}
  version: '1.0'
servers:
- url: https://sfl-clause-analysis-01.herokuapp.com
  variables: {}
paths:
...
  /clause/addClause/100:
    post:
      tags:
      - Clauses
      summary: Add Clause
      operationId: AddClause
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddClauseRequest'
            example:
              createdBy: Arti@21250
              clause: May your work always bring you prosperity.
              context: Social
              visibility: public
        required: true
      responses:
        '200':
          description: 'successful'
          headers: {}
      deprecated: false
      security:
      - apiKey: []
...
  /clause/updateContext/123456/6293c9116e42c551e0f4fd1f:
    put:
      tags:
      - Clauses
      summary: Update Context
      operationId: UpdateContext
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateContextRequest'
            example:
              context: Testing
        required: true
      responses:
        '200':
          description: 'successful'
          headers: {}
      deprecated: false
      security:
      - apiKey: []
  /analysis/getAnalysis:
    post:
      tags:
      - Analysis
      summary: Get A Analysis
      operationId: GetAAnalysis
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GetAAnalysisRequest'
            example:
              clauseID: 6293c9116e42c551e0f4fd1f
        required: true
      responses:
        '200':
          description: 'successful'
          headers: {}
      deprecated: false
      security:
      - apiKey: []
  /analysis/allAnalysis:
    get:
      tags:
      - Analysis
      summary: All Analysis
      operationId: AllAnalysis
      parameters: []
      responses:
        '200':
          description: 'successful'
          headers: {}
      deprecated: false
      security:
      - apiKey: []
  /analysis/treetest:
    post:
      tags:
      - Analysis
      summary: Generate SFL
      operationId: GenerateSFL
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateSFLRequest'
            example:
              body:
              - Clause:
                  ngp:
                    dd:
                      the: {}
                    th:
                      canvas: {}
                    m|adjgp:
                      t:
                        in: {}
                      a:
                        the: {}
                      th:
                        bedroom: {}
        required: true
      responses:
        '200':
          description: 'successful'
          headers: {}
      deprecated: false
      security:
      - apiKey: []
  /analysis/saveAnswers:
    post:
      tags:
      - Analysis
      summary: Save Answers
      operationId: SaveAnswers
      parameters: []
      requestBody:
        description: ''
        content:
          text/plain:
            schema:
              type: string
              example: >-
                {

                    "userName": "Arti@21250",

                    "groupID": 4,

                    "clauseID": "62bdbf6385e9822184c15bd6",

                    "analysis": {{analysis}}

                }
            example: >-
              {

                  "userName": "Arti@21250",

                  "groupID": 4,

                  "clauseID": "62bdbf6385e9822184c15bd6",

                  "analysis": {{analysis}}

              }
        required: true
      responses:
        '200':
          description: 'successful'
          headers: {}
      deprecated: false
      security:
      - apiKey: []
  /analysis/getTree/62d16af07749640004a54293:
    get:
      tags:
      - Analysis
      summary: Get tree
      operationId: Gettree
      parameters: []
      responses:
        '200':
          description: 'successful'
          headers: {}
      deprecated: false
      security:
      - apiKey: []
...
  /analysis/sendInvite:
    post:
      tags:
      - Invite
      summary: send invite
      operationId: sendinvite
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/sendinviterequest'
            example:
              AdminID: arti@071
              GroupID: 123456
              Invite_TO: arti@07
              Status: 'NULL'
        required: true
      responses:
        '200':
          description: 'successful'
          headers: {}
      deprecated: false
      security:
      - apiKey: []
components:
  schemas:
...
    AddClauseRequest:
      title: AddClauseRequest
      required:
      - createdBy
      - clause
      - context
      - visibility
      type: object
      properties:
        createdBy:
          type: string
        clause:
          type: string
        context:
          type: string
        visibility:
          type: string
      example:
        createdBy: Arti@21250
        clause: May your work always bring you prosperity.
        context: Social
        visibility: public
...
    UpdateContextRequest:
      title: UpdateContextRequest
      required:
      - context
      type: object
      properties:
        context:
          type: string
      example:
        context: Testing
    GetAAnalysisRequest:
      title: GetAAnalysisRequest
      required:
      - clauseID
      type: object
      properties:
        clauseID:
          type: string
      example:
        clauseID: 6293c9116e42c551e0f4fd1f
    GenerateSFLRequest:
      title: GenerateSFLRequest
      required:
      - body
      type: object
      properties:
        body:
          type: array
          items:
            $ref: '#/components/schemas/Body'
          description: ''
      example:
        body:
        - Clause:
            ngp:
              dd:
                the: {}
              th:
                canvas: {}
              m|adjgp:
                t:
                  in: {}
                a:
                  the: {}
                th:
                  bedroom: {}
    Body:
      title: Body
      required:
      - Clause
      type: object
      properties:
        Clause:
          $ref: '#/components/schemas/Clause'
      example:
        Clause:
          ngp:
            dd:
              the: {}
            th:
              canvas: {}
            m|adjgp:
              t:
                in: {}
              a:
                the: {}
              th:
                bedroom: {}
    Clause:
      title: Clause
      required:
      - ngp
      type: object
      properties:
        ngp:
          $ref: '#/components/schemas/Ngp'
      example:
        ngp:
          dd:
            the: {}
          th:
            canvas: {}
          m|adjgp:
            t:
              in: {}
            a:
              the: {}
            th:
              bedroom: {}
    Ngp:
      title: Ngp
      required:
      - dd
      - th
      - m|adjgp
      type: object
      properties:
        dd:
          $ref: '#/components/schemas/Dd'
        th:
          $ref: '#/components/schemas/Th'
        m|adjgp:
          $ref: '#/components/schemas/MAdjgp'
      example:
        dd:
          the: {}
        th:
          canvas: {}
        m|adjgp:
          t:
            in: {}
          a:
            the: {}
          th:
            bedroom: {}
    Dd:
      title: Dd
      required:
      - the
      type: object
      properties:
        the:
          type: object
      example:
        the: {}
    Th:
      title: Th
      required:
      - canvas
      type: object
      properties:
        canvas:
          type: object
      example:
        canvas: {}
    MAdjgp:
      title: MAdjgp
      required:
      - t
      - a
      - th
      type: object
      properties:
        t:
          $ref: '#/components/schemas/T'
        a:
          $ref: '#/components/schemas/A'
        th:
          $ref: '#/components/schemas/Th1'
      example:
        t:
          in: {}
        a:
          the: {}
        th:
          bedroom: {}
    T:
      title: T
      required:
      - in
      type: object
      properties:
        in:
          type: object
      example:
        in: {}
    A:
      title: A
      required:
      - the
      type: object
      properties:
        the:
          type: object
      example:
        the: {}
    Th1:
      title: Th1
      required:
      - bedroom
      type: object
      properties:
        bedroom:
          type: object
      example:
        bedroom: {}
    sendinviterequest:
      title: sendinviterequest
      required:
      - AdminID
      - GroupID
      - Invite_TO
      - Status
      type: object
      properties:
        AdminID:
          type: string
        GroupID:
          type: integer
          format: int32
        Invite_TO:
          type: string
        Status:
          type: string
      example:
        AdminID: arti@071
        GroupID: 123456
        Invite_TO: arti@07
        Status: 'NULL'
  securitySchemes:
    apiKey:
      type: apiKey
      name: x-auth-token
      in: header
security: []
...